requirements.txt
streamlit
langgraph
langchain-google-genai
langchain-community
duckduckgo-search
python-dotenv